HW_5/requirements.txt
selenium==4.3.0
lxml==4.9.1
pymongo==4.1.1
python-dotenv==0.20.0
